requirements.txt
Babel==2.9.0
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.7
django-jsonfield==1.4.1
gunicorn==20.1.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
json5==0.9.5
jsonschema==3.2.0
nest-asyncio==1.5.1
pickleshare==0.7.5
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyzmq==22.0.2
requests==2.25.1
six==1.15.0
urllib3==1.26.3
virtualenv==20.4.2
whitenoise==5.2.0

